Redirect all output to log file and show only echo statements
Print only echo statements to the screen. This provides the user with a
clean experience and eliminates noise.

Redirect all other output as well as echo statements' to a log file.
This ensures that the user is able to examine the output at a later
date.

+++ scripts/setup.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+source ${topdir}/scripts/common.sh
+
 function setup_log_file() {
 	# create the log directory if it doesn't already exist
 	LOG_DIR="${topdir}/logs"
@@ -18,10 +20,14 @@ function setup_build_tools() {
         cd ${topdir}/tools/
 
         echo "> Arm Toolchain: not found. Downloading .." 
-        wget https://developer.arm.com/-/media/Files/downloads/gnu-a/9.2-2019.12/binrel/gcc-arm-9.2-2019.12-x86_64-arm-none-linux-gnueabihf.tar.xz
-        tar -Jxf gcc-arm-9.2-2019.12-x86_64-arm-none-linux-gnueabihf.tar.xz
-        rm gcc-arm-9.2-2019.12-x86_64-arm-none-linux-gnueabihf.tar.xz
-        echo "> Arm Toolchain: downloaded .." 
+        wget https://developer.arm.com/-/media/Files/downloads/gnu-a/9.2-2019.12/binrel/gcc-arm-9.2-2019.12-x86_64-arm-none-linux-gnueabihf.tar.xz &>>/dev/null
+		if [ $? -eq 0 ]; then
+			echo "> Arm Toolchain: downloaded .."
+			tar -Jxf gcc-arm-9.2-2019.12-x86_64-arm-none-linux-gnueabihf.tar.xz &>>"${LOG_FILE}"
+			rm gcc-arm-9.2-2019.12-x86_64-arm-none-linux-gnueabihf.tar.xz
+		else
+			echo "> Arm Toolchain: could not download"
+		fi
     else
         echo "> Arm Toolchain: available"
     fi
@@ -33,10 +39,14 @@ function setup_build_tools() {
         cd ${topdir}/tools/
 
         echo "> Aarch64 Toolchain: not found. downloading .." 
-        wget https://developer.arm.com/-/media/Files/downloads/gnu-a/9.2-2019.12/binrel/gcc-arm-9.2-2019.12-x86_64-aarch64-none-linux-gnu.tar.xz
-        tar -Jxf gcc-arm-9.2-2019.12-x86_64-aarch64-none-linux-gnu.tar.xz
-        rm gcc-arm-9.2-2019.12-x86_64-aarch64-none-linux-gnu.tar.xz
-        echo "> Aarch64 Toolchain: downloaded .." 
+        wget https://developer.arm.com/-/media/Files/downloads/gnu-a/9.2-2019.12/binrel/gcc-arm-9.2-2019.12-x86_64-aarch64-none-linux-gnu.tar.xz &>>/dev/null
+		if [ $? -eq 0 ]; then
+			echo "> Aarch64 Toolchain: downloaded .." 
+			tar -Jxf gcc-arm-9.2-2019.12-x86_64-aarch64-none-linux-gnu.tar.xz  &>>"${LOG_FILE}"
+			rm gcc-arm-9.2-2019.12-x86_64-aarch64-none-linux-gnu.tar.xz
+		else
+			echo "> Aarch Toolchain: Could not download"
+		fi
     else
         echo "> Aarch64 Toolchain: available"
     fi
